Fix underflow bug

+++ src/xpcs/main.cpp
@@ -334,7 +334,7 @@ int main(int argc, char** argv)
       // Getting rid of norm factor division for now. 
       //denominator = pixels_per_sbin[i] * swindow * norm_factor;
       
-      denominator = pixels_per_sbin[i] * swindow;
+      denominator = (float)pixels_per_sbin[i] * swindow;
       partial_part_mean[j * total_static_partns + i] /= denominator;
     }
   }
@@ -343,7 +343,7 @@ int main(int argc, char** argv)
     // Getting rid of norm factor division for now. 
     //denominator = pixels_per_sbin[i] * frames * norm_factor;
     
-    denominator = pixels_per_sbin[i] * frames;
+    denominator = (float)pixels_per_sbin[i] * frames;
     partitions_mean[i] /= denominator;
   }
